Minor DB/Test Fixes
- changed generateData to write to the db file instead of command line
- changed generateData to mjs to support newest faker and uuidv4 generation
- deleted old tests (replace later)
